+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 ### Added
+- peer fields: description, address, asn
+
 ### Fixed
 - handle date format for newer bird